GCP_cloud_functions/lake-to-data-warehouse/requirements.txt
google-cloud-storage
google-cloud-pubsub